Cambio tipo de dato

+++ web/MongoProyectoWeb/MongoProyectoWeb/Models/VistaResenasModel.cs
@@ -4,6 +4,6 @@
     {
         public int? _id { get; set; }
         public int? total_resenas { get; set; }
-        public int? promedio_calificacion { get; set; }
+        public float? promedio_calificacion { get; set; }
     }
 }
